Move model constant file to proper package
